Cleaning up temporary files, back-end functions split to each category
